Add motor driver, and telemetry for Flysky receiver.

+++ radio/src/telemetry/telemetry.cpp
@@ -20,7 +20,7 @@
 
 #include "opentx.h"
 
-uint8_t telemetryStreaming = 0;
+uint16_t telemetryStreaming = 0;
 uint8_t telemetryRxBuffer[TELEMETRY_RX_PACKET_SIZE];   // Receive buffer. 9 bytes (full packet), worst case 18 bytes with byte-stuffing (+1)
 uint8_t telemetryRxBufferCount = 0;
